+++ docs/release-notes/change-log.md
@@ -12,6 +12,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 * Max-subrequests can now be overriden by auth header `X-Sf-Substreams-Parallel-Jobs` (note: if your auth plugin is 'trust', make sure that you filter out this header from public access
 * Request Stats logging. When enable it will log metrics associated to a Tier1 and Tier2 request
+* On request, save "substreams.partial.spkg" file to the state cache for debugging purposes.
+* Manifest reader can now read 'partial' spkg files (without protobuf and metadata) with an option.
 
 #### Fixed